daemon: last_activity_epoch_ms crosses the wire as u64, not u128
Measured 2026-09-02 14:0x, first restart onto the fs-truth build: every
`server snapshot` CLIENT parse failed with `u128 is not supported` —
serde_json without `arbitrary_precision` refuses to deserialize u128, and
the field is a WIRE field. Epoch ms fits u64 for the next 500 millennia;
the u128 was borrowed from StartpageDurableRow's internal field and never
belonged on the wire. Daemon fill, both verbs' folds, and the shell's
session conversion widened back to u128 only at the internal row boundary.

+++ crates/yggterm-server/src/lib.rs
@@ -3845,7 +3845,7 @@ pub struct SnapshotSessionView {
     /// honest unknown, never synthesized. serde(default) keeps older
     /// snapshots deserializable.
     #[serde(default)]
-    pub last_activity_epoch_ms: Option<u128>,
+    pub last_activity_epoch_ms: Option<u64>,
     pub ssh_target: Option<String>,
     pub ssh_prefix: Option<String>,
     // Observability (campaign grid-squish diagnosis): the PTY's current grid
@@ -4309,7 +4309,7 @@ pub struct ManagedSessionView {
     /// synthesizes a recency for a row the daemon could not measure
     /// (`None` = unknown, epoch 0, ranked honestly last). 2026-09-02 fs-truth
     /// law: recency outranks live-ness, so the epoch must be real.
-    pub last_activity_epoch_ms: Option<u128>,
+    pub last_activity_epoch_ms: Option<u64>,
     pub ssh_target: Option<String>,
     pub ssh_prefix: Option<String>,
     pub stored_preview_hydrated: bool,
